removed model. refactored mesh primitives into meyer singletons

+++ glSandbox/headers/AssetLoader.h
@@ -1,4 +1,4 @@
 #pragma once
 #include <string>
 
-unsigned int loadGLTF(std::string const& filename);
+size_t loadGLTF(std::string const& filename);
